pyproject.toml
[tool.poetry]
name = "context-data-schema"
version = "0.1.0"
description = "Shared database models and migrations"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = true

packages = [
    { include = "context_db" }
]

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.13.1"
sqlalchemy = "^2.0.29"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
pgvector = "^0.4.2"
geoalchemy2 = "^0.14.6"

[tool.poetry.group.dev.dependencies]
eralchemy2 = "^1.4.1"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
